Move layout-scoped URL helpers from `data.go` to `client.go` for improved organization; no functional changes.
